+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Fixed not being able to use snapshot IDs for --image-id
 - Fixed 404 when waiting for backupunits to be ready
 
+### Changed
+
+- When using config files for authentication, print explicit errors if the file cannot be read
+
 ## [v6.8.0] (January 2025)
 
 ### Added